enh: bring utilities into the new dipy module
